Update README.md
I've changed part of the ``README.md`` to match with the rest of the Karumi projects style. @flipper83, you'll noticed that I've changed part of the ``README.md`` to use ``java`` instead of ``Android``. That's because we said we were going to write this api client without use the Android SDK. Can we also rename the repository from ``MarvelApiClientAndroid`` to ``MarvelApiClientJava``?